+++ src/pipeline_defaults.jl
@@ -1,41 +1,7 @@
-### Global for history, replies, etc
-const CONV_HISTORY = Vector{Vector{PT.AbstractMessage}}()
-const CONV_HISTORY_LOCK = ReentrantLock()
-const MAX_HISTORY_LENGTH = 1
-const LAST_RESULT = Ref{Union{Nothing, RT.AbstractRAGResult}}(nothing)
-const MAIN_INDEX = Ref{Union{Nothing, RT.AbstractChunkIndex}}(nothing)
-"""
-    ALLOWED PACKS
-
-Currently available packs are:
-- `:julia` - Julia documentation, standard library docstrings and a few extras (for Julia v1.10)
-- `:tidier` - Tidier.jl organization documentation (as of 7th April 2024)
-- `:makie` - Makie.jl organization documentation (as of 30th March 2024)
-"""
-const ALLOWED_PACKS = [:julia, :tidier, :makie]
-
-"""
-    LOADED_PACKS
-
-The knowledge packs that are currently loaded in the index.
-"""
-const LOADED_PACKS = Ref{Vector{Symbol}}(@load_preference("LOADED_PACKS",
-    default=["julia"]) .|> Symbol)
-
-### Globals for configuration
-# These serve as reference models to be injected in the absence of inputs, 
-# but the actual used for the query is primarily provided aihelpme directly or via the active RAG_KWARGS
-const MODEL_CHAT = @load_preference("MODEL_CHAT",
-    default="gpt4t")
-const MODEL_EMBEDDING = @load_preference("MODEL_EMBEDDING",
-    default="text-embedding-3-large")
-const EMBEDDING_DIMENSION = @load_preference("EMBEDDING_DIMENSION",
-    default=1024)
-
 # Loaded up with `update_pipeline!` later once RAG CONFIGURATIONS is populated
-const RAG_KWARGS = Ref{NamedTuple}()
-const RAG_CONFIG = Ref{RT.AbstractRAGConfig}()
-const LOADED_CONFIG_KEY = Ref{String}("")  # get the current config key
+global RAG_KWARGS::NamedTuple = NamedTuple()
+global RAG_CONFIG::RT.AbstractRAGConfig = RAGConfig() # just initialize, it will be changed
+global LOADED_CONFIG_KEY::String = "" # get the current config key
 
 """
     RAG_CONFIGURATIONS
@@ -49,8 +15,12 @@ Available Options:
 - `:silver`: A simple configuration for a bronze pipeline, using truncated binary embeddings (dimensionality: 1024) but also enables re-ranking step.
 - `:gold`: A more complex configuration, similar to `:simpler`, but using a standard embeddings (dimensionality: 3072, type: Float32). It also leverages re-ranking and refinement with a web-search.
 """
-const RAG_CONFIGURATIONS = let MODEL_CHAT = MODEL_CHAT, MODEL_EMBEDDING = MODEL_EMBEDDING
-    RAG_CONFIGURATIONS = Dict{Symbol, Dict{Symbol, Any}}()
+global RAG_CONFIGURATIONS::Dict{Symbol, Dict{Symbol, Any}} = Dict{
+    Symbol, Dict{Symbol, Any}}()
+
+## Load configurations
+let MODEL_CHAT = MODEL_CHAT, MODEL_EMBEDDING = MODEL_EMBEDDING,
+    RAG_CONFIGURATIONS = RAG_CONFIGURATIONS
     ## Bronze
     RAG_CONFIGURATIONS[:bronze] = Dict{Symbol, Any}(
         :config => RT.RAGConfig(;
@@ -70,9 +40,6 @@ const RAG_CONFIGURATIONS = let MODEL_CHAT = MODEL_CHAT, MODEL_EMBEDDING = MODEL_
             generator_kwargs = (;
                 answerer_kwargs = (;
                     model = MODEL_CHAT),
-                embedder_kwargs = (;
-                    truncate_dimension = EMBEDDING_DIMENSION,
-                    model = MODEL_EMBEDDING),
                 refiner_kwargs = (;
                     model = MODEL_CHAT))))
     ## Silver - reranking added
@@ -95,9 +62,6 @@ const RAG_CONFIGURATIONS = let MODEL_CHAT = MODEL_CHAT, MODEL_EMBEDDING = MODEL_
             generator_kwargs = (;
                 answerer_kwargs = (;
                     model = MODEL_CHAT),
-                embedder_kwargs = (;
-                    truncate_dimension = EMBEDDING_DIMENSION,
-                    model = MODEL_EMBEDDING),
                 refiner_kwargs = (;
                     model = MODEL_CHAT))))
     ## Gold  - reranking + web-search
@@ -114,23 +78,20 @@ const RAG_CONFIGURATIONS = let MODEL_CHAT = MODEL_CHAT, MODEL_EMBEDDING = MODEL_
                 rephraser_kwargs = (;
                     model = MODEL_CHAT),
                 embedder_kwargs = (;
+                    truncate_dimension = nothing,
                     model = MODEL_EMBEDDING),
                 tagger_kwargs = (;
                     model = MODEL_CHAT)),
             generator_kwargs = (;
                 answerer_kwargs = (;
                     model = MODEL_CHAT),
-                embedder_kwargs = (;
-                    model = MODEL_EMBEDDING),
                 refiner_kwargs = (;
                     model = MODEL_CHAT))))
-
-    RAG_CONFIGURATIONS
 end
 
 "Returns the configuration key for the given `cfg` and `kwargs` to use the relevant artifacts."
 function get_config_key(
-        cfg::AbstractRAGConfig = RAG_CONFIG[], kwargs::NamedTuple = RAG_KWARGS[])
+        cfg::AbstractRAGConfig = RAG_CONFIG, kwargs::NamedTuple = RAG_KWARGS)
     emb_model = getpropertynested(kwargs, [:embedder_kwargs], :model)
     emb_dim = getpropertynested(kwargs, [:embedder_kwargs], :truncate_dimension, 0)
     emb_eltype = RT.EmbedderEltype(cfg.retriever.embedder)
@@ -151,7 +112,7 @@ See available pipeline options via `keys(RAG_CONFIGURATIONS)`.
 
 Logic:
 - Updates the global `MODEL_CHAT` and `MODEL_EMBEDDING` to the requested models.
-- Update the global `EMBEDDING_DIMENSION` for the requested embedding dimensionality after truncation (`embedding_dimension`).
+- Updates the global `EMBEDDING_DIMENSION` for the requested embedding dimensionality after truncation (`embedding_dimension`).
 - Updates the global `RAG_CONFIG` and `RAG_KWARGS` to the requested `option`.
 - Updates the global `LOADED_CONFIG_KEY` to the configuration key for the given `option` and `kwargs` (used by the artifact system to download the correct knowledge packs).
 
@@ -171,25 +132,33 @@ update_pipeline!(:bronze; model_chat = "llama3", model_embedding="nomic-embed-te
 load_index!()
 ```
 """
-function update_pipeline!(option::Symbol = :bronze; model_chat = MODEL_CHAT,
-        model_embedding = MODEL_EMBEDDING, verbose::Bool = true,
-        embedding_dimension::Integer = EMBEDDING_DIMENSION)
+function update_pipeline!(
+        option::Symbol = :bronze; model_chat::Union{String, Nothing} = nothing,
+        model_embedding::Union{String, Nothing} = nothing, verbose::Bool = true,
+        embedding_dimension::Union{Integer, Nothing} = nothing)
     global RAG_CONFIGURATIONS, RAG_CONFIG, RAG_KWARGS, MODEL_CHAT, MODEL_EMBEDDING, EMBEDDING_DIMENSION, LOADED_CONFIG_KEY
 
-    @assert haskey(RAG_CONFIGURATIONS, option) "Invalid option: $option. Select one of: $(join(keys(RAG_CONFIGURATIONS),", "))"
-    @assert embedding_dimension in [0, 1024, 3072] "Invalid embedding_dimension: $(embedding_dimension). Supported: 0, 1024, 3072. See the available artifacts."
-    ## Model-specific checks, they do not fail but at least warn
-    if model_embedding == "nomic-embed-text" && !iszero(embedding_dimension)
-        @warn "Invalid configuration for knowledge packs! For `nomic-embed-text`, `embedding_dimension` must be 0. See the available artifacts."
-    end
-    if model_embedding == "text-embedding-3-large" && embedding_dimension ∉ [1024, 0]
-        @warn "Invalid configuration for knowledge packs! For `text-embedding-3-large`, `embedding_dimension` must be 0 or 1024. See the available artifacts."
+    ## Set from globals if not provided
+    model_chat = !isnothing(model_chat) ? model_chat : MODEL_CHAT
+    model_embedding = !isnothing(model_embedding) ? model_embedding : MODEL_EMBEDDING
+    ## Do not set embedding dimensions, we might need to extract it from kwargs
+
+    ## WARN about limited support for nomic-embed-text -- we need to create repeatable process
+    if model_embedding == "nomic-embed-text"
+        @warn "Knowledge packs for `nomic-embed-text` are currently not built automatically, so they might be missing / outdated. Please switch to OpenAI `text-embedding-3-large` for the best experience."
     end
 
-    ## Update model references
-    MODEL_CHAT = model_chat
-    MODEL_EMBEDDING = model_embedding
-    EMBEDDING_DIMENSION = embedding_dimension
+    @assert haskey(RAG_CONFIGURATIONS, option) "Invalid option: $option. Select one of: $(join(keys(RAG_CONFIGURATIONS),", "))"
+    @assert (isnothing(embedding_dimension)||embedding_dimension in [0, 1024]) "Invalid embedding_dimension: $(embedding_dimension). Supported: 0 (no truncation) or 1024. See the available artifacts."
+    ## Model-specific checks, they do not fail but at least warn
+    if model_embedding == "nomic-embed-text" &&
+       !(iszero(embedding_dimension) || isnothing(embedding_dimension))
+        @warn "Invalid configuration for knowledge packs! For `nomic-embed-text`, `embedding_dimension` must be 0. See the available artifacts."
+    end
+    if model_embedding == "text-embedding-3-large" &&
+       (embedding_dimension ∉ [1024, 0] || !isnothing(embedding_dimension))
+        @warn "Invalid configuration for knowledge packs! For `text-embedding-3-large`, `embedding_dimension` must be 0 or 1024. See the available artifacts."
+    end
 
     config = RAG_CONFIGURATIONS[option][:config]
     kwargs = RAG_CONFIGURATIONS[option][:kwargs]
@@ -206,16 +175,26 @@ function update_pipeline!(option::Symbol = :bronze; model_chat = MODEL_CHAT,
     if !isnothing(embedding_dimension)
         kwargs = setpropertynested(
             kwargs, [:embedder_kwargs], :truncate_dimension, embedding_dimension)
+    else
+        ## load the value from defaults in config -- to match the artifacts
+        val = getpropertynested(kwargs, [:embedder_kwargs], :truncate_dimension, nothing)
+        embedding_dimension = isnothing(val) ? 0 : val
     end
+
+    ## Update GLOBAL variables
+    MODEL_CHAT = model_chat
+    MODEL_EMBEDDING = model_embedding
+    @info embedding_dimension
+    EMBEDDING_DIMENSION = embedding_dimension
 
     ## Set the options
     config_key = get_config_key(config, kwargs)
     ## detect significant changes
-    !isempty(LOADED_CONFIG_KEY[]) && LOADED_CONFIG_KEY[] != config_key &&
+    !isempty(LOADED_CONFIG_KEY) && LOADED_CONFIG_KEY != config_key &&
         @warn "Core RAG pipeline configuration has changed! You must re-build your index with `AIHelpMe.load_index!()`!"
-    LOADED_CONFIG_KEY[] = config_key
-    RAG_KWARGS[] = kwargs
-    RAG_CONFIG[] = config
+    LOADED_CONFIG_KEY = config_key
+    RAG_KWARGS = kwargs
+    RAG_CONFIG = config
 
     verbose &&
         @info "Updated RAG pipeline to `:$option` (Configuration key: \"$config_key\")."